+++ lib/src/api/bridge/frb.dart
@@ -12,6 +12,7 @@ export 'api/media_stream_track/media_track_state.dart';
 export 'api/rtc_ice_candidate_stats.dart';
 export 'api/rtc_rtp_encoding_parameters.dart';
 export 'api/rtc_rtp_send_parameters.dart';
+export 'api/rtc_rtp_transceiver.dart';
 export 'api/rtc_session_description.dart';
 export 'api/rtc_stats.dart';
 export 'api/rtc_stats/ice_role.dart';
